ProjectUICommands: mv prepareForExit method
