+++ src/org/omegat/gui/main/ProjectUICommands.java
@@ -53,6 +53,7 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang.builder.EqualsBuilder;
 
 import org.omegat.CLIParameters;
+import org.omegat.Main;
 import org.omegat.convert.ConvertProject;
 import org.omegat.core.Core;
 import org.omegat.core.CoreEvents;
@@ -62,14 +63,17 @@ import org.omegat.core.data.ProjectProperties;
 import org.omegat.core.events.IProjectEventListener;
 import org.omegat.core.segmentation.SRX;
 import org.omegat.core.segmentation.Segmenter;
+import org.omegat.core.spellchecker.ISpellChecker;
 import org.omegat.core.team2.IRemoteRepository2;
 import org.omegat.core.team2.RemoteRepositoryProvider;
 import org.omegat.filters2.master.FilterMaster;
+import org.omegat.filters2.master.PluginUtils;
 import org.omegat.gui.dialogs.ChooseMedProject;
 import org.omegat.gui.dialogs.FileCollisionDialog;
 import org.omegat.gui.dialogs.NewProjectFileChooser;
 import org.omegat.gui.dialogs.NewTeamProjectController;
 import org.omegat.gui.dialogs.ProjectPropertiesDialog;
+import org.omegat.gui.editor.SegmentExportImport;
 import org.omegat.util.FileUtil;
 import org.omegat.util.FileUtil.ICollisionCallback;
 import org.omegat.util.HttpConnectionUtils;
@@ -1307,6 +1311,87 @@ public final class ProjectUICommands {
             Log.log(ex);
             Core.getMainWindow().displayErrorRB(ex, "TF_WIKI_IMPORT_FAILED");
         }
+    }
+
+    public static void saveAndExit() {
+        saveAndExitOrRestart(false, null);
+    }
+
+    public static void saveAndRestart(String projectDir) {
+        saveAndExitOrRestart(true, projectDir);
+    }
+
+    /**
+     * Quit or restart OmegaT.
+     */
+    private static void saveAndExitOrRestart(boolean restart, String projectDir) {
+        // Bug #902: commit the current entry first
+        // We do it before checking project status, so that it can eventually change it
+        if (Core.getProject().isProjectLoaded()) {
+            Core.getEditor().commitAndLeave();
+        }
+
+        boolean projectModified = false;
+        if (Core.getProject().isProjectLoaded()) {
+            projectModified = Core.getProject().isProjectModified();
+        }
+        // RFE 1302358
+        // Add Yes/No Warning before OmegaT quits
+        if (projectModified || Preferences.isPreference(Preferences.ALWAYS_CONFIRM_QUIT)) {
+            if (JOptionPane.YES_OPTION != JOptionPane.showConfirmDialog(
+                    Core.getMainWindow().getApplicationFrame(),
+                    OStrings.getString("MW_QUIT_CONFIRM"), OStrings.getString("CONFIRM_DIALOG_TITLE"),
+                    JOptionPane.YES_NO_OPTION)) {
+                return;
+            }
+        }
+
+        SegmentExportImport.flushExportedSegments();
+
+        new SwingWorker<Void, Void>() {
+            @Override
+            protected Void doInBackground() throws Exception {
+                if (Core.getProject().isProjectLoaded()) {
+                    // Save the list of learned and ignore words
+                    ISpellChecker sc = Core.getSpellChecker();
+                    sc.saveWordLists();
+                    try {
+                        Core.executeExclusively(true, () -> {
+                            Core.getProject().saveProject(true);
+                            ProjectFactory.closeProject();
+                        });
+                    } catch (KnownException ex) {
+                        // hide exception on shutdown
+                    }
+                }
+
+                CoreEvents.fireApplicationShutdown();
+
+                PluginUtils.unloadPlugins();
+
+                return null;
+            }
+
+            @Override
+            protected void done() {
+                try {
+                    get();
+
+                    MainWindowUI.saveScreenLayout();
+
+                    Preferences.save();
+
+                    if (restart) {
+                        Main.restartGUI(projectDir);
+                    } else {
+                        System.exit(0);
+                    }
+                } catch (Exception ex) {
+                    Log.logErrorRB(ex, "PP_ERROR_UNABLE_TO_READ_PROJECT_FILE");
+                    Core.getMainWindow().displayErrorRB(ex, "PP_ERROR_UNABLE_TO_READ_PROJECT_FILE");
+                }
+            }
+        }.execute();
     }
 
     private static boolean ensureProjectDir(File dir) {
